Moved all frontend files to a package "frontend"
full package name is: `com.agora.app.frontend`
