+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.AppCompat.Light"
         tools:targetApi="31">
         <activity
-            android:name=".RegisterActivity"
+            android:name=".frontend.RegisterActivity"
             android:exported="true"
             android:label="@string/app_name"
             android:theme="@style/Theme.Agora">
@@ -23,12 +23,12 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".MarketplaceActivity"/>
-        <activity android:name=".SwipingActivity"/>
-        <activity android:name=".UserInfoActivity"/>
-        <activity android:name=".AccountCreatorActivity"/>
-        <activity android:name=".LoginActivity"/>
-      
+        <activity android:name=".frontend.MarketplaceActivity"/>
+        <activity android:name=".frontend.SwipingActivity"/>
+        <activity android:name=".frontend.UserInfoActivity"/>
+        <activity android:name=".frontend.AccountCreatorActivity"/>
+        <activity android:name=".frontend.LoginActivity"/>
+
     </application>
 
 </manifest>
