+++ app/src/main/java/com/agora/app/frontend/AccountCreatorActivity.java
@@ -1,4 +1,4 @@
-package com.agora.app;
+package com.agora.app.frontend;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -7,6 +7,7 @@ import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.Toast;
+import com.agora.app.R;
 
 public class AccountCreatorActivity extends AppCompatActivity {
 
@@ -33,12 +34,12 @@ public class AccountCreatorActivity extends AppCompatActivity {
                 // Check if username or password fields are empty
                 if (username.isEmpty()) {
                     Toast.makeText(AccountCreatorActivity.this, "Username cannot be empty", Toast.LENGTH_SHORT).show();
-                    return;  
+                    return;
                 }
 
                 if (password.isEmpty()) {
                     Toast.makeText(AccountCreatorActivity.this, "Password cannot be empty", Toast.LENGTH_SHORT).show();
-                    return;  
+                    return;
                 }
 
                 if (password.equals(confirm_pass.getText().toString())) {
