+++ app/src/main/java/com/agora/app/frontend/SwipingActivity.java
@@ -1,10 +1,10 @@
-package com.agora.app;
+package com.agora.app.frontend;
 
 import android.content.Intent;
 import android.os.Bundle;
 import androidx.appcompat.app.AppCompatActivity;
-import android.view.View;
 import android.widget.Button;
+import com.agora.app.R;
 
 public class SwipingActivity extends AppCompatActivity {
     @Override
